mvp-v2 feat: MVP v2 백엔드 계약 반영

+++ src/main/java/com/saynow/practice/api/dto/MicReadyResponse.java
@@ -2,6 +2,6 @@ package com.saynow.practice.api.dto;
 
 public record MicReadyResponse(
         String sessionId,
-        long micReadyLatencyMs
+        Integer micReadyLatencyMs
 ) {
 }
